+++ packages/components/modal/src/index.ts
@@ -1,5 +1,4 @@
 import Modal from "./modal";
-import ModalTrigger from "./modal-trigger";
 import ModalContent from "./modal-content";
 import ModalHeader from "./modal-header";
 import ModalBody from "./modal-body";
@@ -7,7 +6,6 @@ import ModalFooter from "./modal-footer";
 
 // export types
 export type {ModalProps} from "./modal";
-export type {ModalTriggerProps} from "./modal-trigger";
 export type {ModalContentProps} from "./modal-content";
 export type {ModalHeaderProps} from "./modal-header";
 export type {ModalBodyProps} from "./modal-body";
@@ -15,9 +13,10 @@ export type {ModalFooterProps} from "./modal-footer";
 
 // export hooks
 export {useModal} from "./use-modal";
+export * from "@nextui-org/use-disclosure";
 
 // export context
 export * from "./modal-context";
 
 // export components
-export {Modal, ModalTrigger, ModalContent, ModalHeader, ModalBody, ModalFooter};
+export {Modal, ModalContent, ModalHeader, ModalBody, ModalFooter};