+++ packages/components/dropdown/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nextui-org/dropdown
 
+## 0.0.0-dev-v2-20230420134926
+
+### Patch Changes
+
+- Modal structure change, now is controlled by the useDisclosure modal
+- Updated dependencies
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230420134926
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230420134926
+  - @nextui-org/use-is-mobile@0.0.0-dev-v2-20230420134926
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230420134926
+  - @nextui-org/popover@0.0.0-dev-v2-20230420134926
+  - @nextui-org/system@0.0.0-dev-v2-20230420134926
+  - @nextui-org/theme@0.0.0-dev-v2-20230420134926
+
 ## 0.0.0-dev-v2-20230420024722
 
 ### Patch Changes