+++ packages/components/navbar/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nextui-org/navbar
 
+## 0.0.0-dev-v2-20230420134926
+
+### Patch Changes
+
+- Modal structure change, now is controlled by the useDisclosure modal
+- Updated dependencies
+  - @nextui-org/use-aria-toggle-button@0.0.0-dev-v2-20230420134926
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230420134926
+  - @nextui-org/use-scroll-position@0.0.0-dev-v2-20230420134926
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230420134926
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230420134926
+  - @nextui-org/system@0.0.0-dev-v2-20230420134926
+  - @nextui-org/theme@0.0.0-dev-v2-20230420134926
+
 ## 0.0.0-dev-v2-20230420024722
 
 ### Patch Changes