+++ packages/core/react/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @nextui-org/react
 
+## 0.0.0-dev-v2-20230420135820
+
+### Patch Changes
+
+- Updated dependencies
+  - @nextui-org/modal@0.0.0-dev-v2-20230420135820
+
+## 0.0.0-dev-v2-20230420134926
+
+### Patch Changes
+
+- Modal structure change, now is controlled by the useDisclosure modal
+- Updated dependencies
+  - @nextui-org/pagination@0.0.0-dev-v2-20230420134926
+  - @nextui-org/accordion@0.0.0-dev-v2-20230420134926
+  - @nextui-org/checkbox@0.0.0-dev-v2-20230420134926
+  - @nextui-org/dropdown@0.0.0-dev-v2-20230420134926
+  - @nextui-org/progress@0.0.0-dev-v2-20230420134926
+  - @nextui-org/popover@0.0.0-dev-v2-20230420134926
+  - @nextui-org/snippet@0.0.0-dev-v2-20230420134926
+  - @nextui-org/spinner@0.0.0-dev-v2-20230420134926
+  - @nextui-org/tooltip@0.0.0-dev-v2-20230420134926
+  - @nextui-org/avatar@0.0.0-dev-v2-20230420134926
+  - @nextui-org/button@0.0.0-dev-v2-20230420134926
+  - @nextui-org/navbar@0.0.0-dev-v2-20230420134926
+  - @nextui-org/switch@0.0.0-dev-v2-20230420134926
+  - @nextui-org/badge@0.0.0-dev-v2-20230420134926
+  - @nextui-org/image@0.0.0-dev-v2-20230420134926
+  - @nextui-org/input@0.0.0-dev-v2-20230420134926
+  - @nextui-org/modal@0.0.0-dev-v2-20230420134926
+  - @nextui-org/radio@0.0.0-dev-v2-20230420134926
+  - @nextui-org/card@0.0.0-dev-v2-20230420134926
+  - @nextui-org/chip@0.0.0-dev-v2-20230420134926
+  - @nextui-org/code@0.0.0-dev-v2-20230420134926
+  - @nextui-org/drip@0.0.0-dev-v2-20230420134926
+  - @nextui-org/link@0.0.0-dev-v2-20230420134926
+  - @nextui-org/user@0.0.0-dev-v2-20230420134926
+  - @nextui-org/system@0.0.0-dev-v2-20230420134926
+  - @nextui-org/theme@0.0.0-dev-v2-20230420134926
+
 ## 0.0.0-dev-v2-20230420024722
 
 ### Patch Changes