+++ packages/hooks/use-safe-layout-effect/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nextui-org/use-safe-layout-effect
 
+## 0.0.0-dev-v2-20230420134926
+
+### Patch Changes
+
+- Modal structure change, now is controlled by the useDisclosure modal
+
 ## 0.0.0-dev-v2-20230420024722
 
 ### Patch Changes